Query the Travis CI API
Query the given resource of the Travis CI API using the access token
found in '~/.travis/config.yml', requesting 20 entries, because that's
convenient for me, and eager-load the information about the jobs, to
avoid individual API requests for each job.

Technically, the repo-slug in the query URL must be properly
urlencoded.  In practice, however:

  - GitHub won't allow such characters in user or repository names
    that need to be urlencoded in the first place, so the only such
    character will be the '/' in between "user/repo".

  - Proper urlencoding is done with 'urllib' in Python, which went
    through some significant API changes between Python 2.x and 3.x.

Because of the above, let's just punt on proper urlencoding, and
simpy do a string replace to turn '/' into "%2F".