Print status in color
The colors mostly match the colors used on the Travis CI website, i.e.
successful build jobs are shown in green, failed or errored in red, and
created in yellow.  The exceptions are already booting and running
build jobs, as they will be shown in blue to better stand out from the
others; on Travis CI they are yellow, too, with a little spinning
animation.  Cancelled builds are left without color; there is not much
to color on an '_' character anyway.

Sensibly, only do coloring when stdout is attached to a tty.

Won't work on Windows, unless the user configured the terminal to
understand ANSI color sequences.