Add '-n|--count <N>' option to specify the number of builds to show
Querying the Travis CI API and retrieving the information of 20 builds
with eager-loaded details about their jobs tends to take a couple of
seconds.

Add the '--count <N>' option to override this default.  Limit it to a
maximum of 100 builds, because that's the max allowed by the Travis CI
API, is plenty for all my use cases, and I would punt on dealing with
pagination at the moment.