Include the build number in the output
The Travis CI website doesn't include the branch name in the title of
the page, so it's not visible in the browser's tab bar, making it
difficult to find the tab corresponding to one of the builds, when
multiple build tabs are open.

However, the title includes the build number, so let's include that in
the compact status as well.